Move Tooltip classes to a namespace
