+++ MonkeyLoader.Resonite.Integration/UI/Tooltips/ResolveTooltipLabelEvent.cs
